+++ src/App.js
@@ -1,8 +1,7 @@
 import React from "react";
 import "./index.css";
-import { Login, Security, Themes, Upgrade, Signup } from "./pages";
-import UserProvider from "./contexts/AuthContext";
-import TaskContextProvider from "./contexts/TaskContext";
+import { Login, Reset, Signup } from "./pages";
+import { Inbox, Security, Themes, Profile, Upgrade } from "./pages";
 import {
   BrowserRouter as Router,
   Switch,
@@ -10,11 +9,11 @@ import {
   Redirect,
 } from "react-router-dom";
 import PrivateRoute from "./PrivateRoute";
-import Reset from "./pages/Reset";
-import Profile from "./pages/Profile";
-import EntryContextProvider from "./contexts/EntryContext";
-import Settings from "./pages/Settings";
-import Inbox from "./pages/Inbox";
+import {
+  TaskContextProvider,
+  EntryContextProvider,
+  UserProvider,
+} from "./contexts";
 
 function App() {
   return (
@@ -33,7 +32,6 @@ function App() {
               <PrivateRoute path="/security" component={Security} />
               <PrivateRoute path="/themes" component={Themes} />
               <PrivateRoute path="/upgrade" component={Upgrade} />
-              <PrivateRoute path="/settings" component={Settings} />
               <Route path="/login" component={Login} />
               <Route path="/signup" component={Signup} />
               <Route path="/reset" component={Reset} />
