+++ src/components/Body/Body.js
@@ -1,5 +1,10 @@
 import React from "react";
-import { BodyContainer } from "../../pages/styles";
+import styled from "styled-components";
+
+export const BodyContainer = styled.div`
+  display: flex;
+  height: 90%;
+`;
 
 const Body = ({ children }) => {
   return <BodyContainer>{children}</BodyContainer>;
