+++ src/components/Body/Content.js
@@ -1,5 +1,11 @@
 import React from "react";
-import { ContentContainer } from "../../pages/styles";
+import styled from "styled-components";
+
+export const ContentContainer = styled.div`
+  width: 80%;
+  height: 100%;
+  position: relative;
+`;
 
 const Content = ({ children }) => {
   return <ContentContainer>{children}</ContentContainer>;
