+++ src/components/Body/index.js
@@ -1,3 +1,4 @@
 export { default as Body } from "./Body";
 export { default as Content } from "./Content";
 export { default as SideBar } from "./SideBar";
+export { default as UserProfilePic } from "./UserProfilePic";
