+++ src/components/Header/Header.js
@@ -1,5 +1,13 @@
 import React from "react";
-import { HeaderContainer } from "../../pages/styles";
+import styled from "styled-components";
+
+export const HeaderContainer = styled.div`
+  height: 10%;
+  width: 100wh;
+  background-color: #efefef;
+  display: flex;
+  align-items: center;
+`;
 
 const Header = ({ children }) => {
   return <HeaderContainer>{children}</HeaderContainer>;
