--- src/components/Task/index.js
@@ -1,1 +1,0 @@
-export { default as Task } from "./Task";
