--- src/components/TaskInput/index.js
@@ -1,1 +1,0 @@
-export { default as TaskInput } from "./TaskInput";
