+++ src/components/TaskList/index.js
@@ -1,1 +1,3 @@
 export { default as TaskList } from "./TaskList";
+export { default as Task } from "../TaskList/Task";
+export { default as TaskInput } from "./TaskInput";
