--- src/components/UserProfilePic.js
@@ -1,13 +1,0 @@
-import React from "react";
-import { RiUserLine } from "react-icons/ri";
-import { UserProfilePicWrapper } from "../pages/styles";
-
-const UserProfilePic = () => {
-  return (
-    <UserProfilePicWrapper>
-      <RiUserLine />
-    </UserProfilePicWrapper>
-  );
-};
-
-export default UserProfilePic;
