--- src/components/index.js
@@ -1,3 +1,0 @@
-export { default as Search } from "./Header/Search";
-
-export { default as SideBar } from "./Body/SideBar";
