+++ src/pages/index.js
@@ -1,5 +1,8 @@
+export { default as Inbox } from "./Inbox";
 export { default as Login } from "./Login";
 export { default as Security } from "../pages/Security";
 export { default as Themes } from "../pages/Themes";
 export { default as Upgrade } from "../pages/Upgrade";
 export { default as Signup } from "./Signup";
+export { default as Profile } from "./Profile";
+export { default as Reset } from "./Reset";
